reorganize plot methods
