+++ src/micromanager_gui/_plate_viewer/_plot_methods/__init__.py
@@ -1,0 +1,3 @@
+from ._main_plot import plot_multi_well_data, plot_single_well_data
+
+__all__ = ["plot_multi_well_data", "plot_single_well_data"]
